Manual schema definition
The idea is to allow to define the underlying schema in several
different ways:

- Manually configuring it on each POST request.
- Set it up through configuration files.
- Set it up from arbitrary endpoint.

TODO:
- [ ] Some refactor to better express separation of concerns
- [ ] Allow defining schemas on same POST request
- [ ] Allow to load configuration files with schema definitions
