+++ src/executor/database/sqlite_extensions/math.so
