+++ src/executor/database/sqlite_extensions/text.so
