+++ src/executor/database/sqlite_extensions/vsv.so
